Move `TokenMap` type out of `TokenBundle` module hierarchy.
The `TokenMap` type, apart from being a sub-component of `TokenBundle`,
is useful in its own right.

When the wallet eventually adds support for the `mint` field of a
transaction (which will encode tokens the transaction is putting into or
taking out of circulation), the wallet will need a way to represent
token bundles that cannot hold ada quantities.

The `TokenMap` type satisfies this use case, as it can only hold non-ada
token quantities.